Verifica se coluna existe antes de excluir

+++ database/migrations/remove/2022_12_09_091721_remove_field_in_table_escola.php
@@ -9,8 +9,11 @@ return new class extends Migration
     public function up()
     {
         DB::unprepared('ALTER TABLE IF EXISTS pmieducar.escola DROP CONSTRAINT IF EXISTS escola_ref_cod_escola_rede_ensino_fkey;');
-        Schema::table('pmieducar.escola', function (Blueprint $table) {
-            $table->dropColumn('ref_cod_escola_rede_ensino');
-        });
+
+        if (Schema::hasColumn('pmieducar.escola', 'ref_cod_escola_rede_ensino')) {
+            Schema::table('pmieducar.escola', function (Blueprint $table) {
+                $table->dropColumn('ref_cod_escola_rede_ensino');
+            });
+        }
     }
 };
